cilium: bpf, use mcpu=v2 to avoid alu32 ops on older kernels
 [ upstream commit 809bfff5678996dc0affc98874b0de3c6294e314 ]

Done independent of upstream commit but more or less the same idea as
described in detail with above commit.

Older kernels had trouble tracking bounds on alu32 operations, but if
we leave mcpu unspecified and probe we may end up running mcpu=v3 which
will let verifier use the alu32 ops.

This can result in errors when loading bpf programs. For backports we
don't want to rework bpf programs to try and get alu32 bit code that
works with verifier. Its not even clear, to me at least, this is even
easily doable on master. So for now on v1.7 branch set mcpu=v2.

Upstream commit,

 commit 809bfff5678996dc0affc98874b0de3c6294e314
 Author: Daniel Borkmann <daniel@iogearbox.net>
 Date:   Mon Mar 23 12:06:51 2020 +0100

    bpf, cilium: manually probe for v1/v2 BPF CPU availability

does a bit more probing before setting it directly here we wont bother
to push through golang side and hard code in init.sh to minimize changes.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>